+++ src/client/screens/OrganizationAdminScreen.tsx
@@ -1,19 +1,22 @@
 /** @jsx jsx */
 import { useEffect } from "react";
 import { connect } from "react-redux";
-import { useParams, Link } from "react-router-dom";
-import { Box, Flex, Heading, jsx, Button } from "theme-ui";
+import { Link, useParams } from "react-router-dom";
+import { Box, Flex, Heading, jsx } from "theme-ui";
+
 import { organizationFetch } from "../actions/organization";
-import { organizationProjectsFetch, exportProjects } from "../actions/organizationProjects";
+import { organizationProjectsFetch } from "../actions/organizationProjects";
+import { userFetch } from "../actions/user";
+import OrganizationAdminProjectsTable from "../components/OrganizationAdminProjectsTable";
+import OrganizationExportMenu from "../components/OrganizationExportMenu";
+import SiteHeader from "../components/SiteHeader";
 import { State } from "../reducers";
 import { OrganizationState } from "../reducers/organization";
+import { OrganizationProjectsState } from "../reducers/organizationProjects";
 import { UserState } from "../reducers/user";
 import store from "../store";
-import SiteHeader from "../components/SiteHeader";
+
 import PageNotFoundScreen from "./PageNotFoundScreen";
-import { OrganizationProjectsState } from "../reducers/organizationProjects";
-import OrganizationAdminProjectsTable from "../components/OrganizationAdminProjectsTable";
-import { userFetch } from "../actions/user";
 
 interface StateProps {
   readonly organization: OrganizationState;
@@ -118,15 +121,10 @@ const OrganizationAdminScreen = ({ organization, user, organizationProjects }: S
                   <Link to={`/o/${organizationSlug}`}>{organization.resource.name}</Link>
                 </Heading>
                 <Heading>Maps</Heading>
-                <Button
-                  onClick={() => store.dispatch(exportProjects(organizationSlug))}
-                  sx={{ float: "right" }}
-                >
-                  Export maps
-                </Button>
                 <Box>
                   Published maps that were created by members of your organization. You can select
                   up to 12 maps to feature on your organization profile page.
+                  <OrganizationExportMenu slug={organizationSlug} />
                 </Box>
               </Box>
             </Flex>